pyproject.toml
[tool.poetry]
name = "calitp"
version = "2022.10.27"
description = "Shared code for the Cal-ITP data codebases"
authors = ["Andrew Vaccaro <atvaccaro@gmail.com>"]
license = "GNU GENERAL PUBLIC LICENSE Version 3, 29 June 2007"

[tool.poetry.dependencies]
python = ">=3.8,<3.11"
fsspec = "^2022.5.0,!=2022.7.1" # https://github.com/fsspec/gcsfs/issues/486
gcsfs = "^2022.5.0,!=2022.7.1" # https://github.com/fsspec/gcsfs/issues/486
pandas = "^1.3.3"
sqlalchemy-bigquery = "^1.4.4"
google-cloud-bigquery = "<3.0.0dev,>=1.15.0"
gtfs-realtime-bindings = "^0.0.7"
humanize = "^4.2.3"
pendulum = "^2.1.2"
pydantic = "^1.9.1"
siuba = "^0.4.0"
Jinja2 = "<3.1.0"
# These are just to play nicely with composer
google-cloud-bigquery-storage = "2.14.1"
google-api-core = "<2.0.0dev,>=1.32.0"
protobuf = ">=3.19.0,<4.0.0dev"
tqdm = "^4.64.0"
backoff = "^2.1.2"
google-cloud-secret-manager = "^2.12.4"
pandas-gbq = "^0.17.9"

[tool.poetry.dev-dependencies]
black = "^22.6.0"
pytest = "^7.1.2"
hypothesis = "^6.54.1"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 120
